% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitespace.R
\name{whitespaced}
\alias{whitespaced}
\title{A special whitespace class that fails on whitespace last}
\usage{
whitespaced(x, before = whitespace(NA, NA), after = whitespace(NA, NA))
}
\description{
A helpful lazy whitespace checker to prioritize syntactic checks over
stylistic checks.
}
